[C] Organize; add display names + prop types
