--- client/src/frontend/components/privacy/CookiesForm/RadioGroup/index.js
@@ -1,1 +1,0 @@
-export { default } from "./RadioGroup";
