--- client/src/global/components/sign-in-up/BaseHookForm/index.js
@@ -1,1 +1,0 @@
-export { default } from "./BaseHookForm";
