+++ client/src/global/components/sign-in-up/EditProfileForm/Greeting/index.js
@@ -1,1 +1,37 @@
-export { default } from "./Greeting";
+import React from "react";
+import { Trans } from "react-i18next";
+import { useFormContext } from "react-hook-form";
+import PropTypes from "prop-types";
+import * as Styled from "./styles";
+
+export default function ProfileGreeting({ mode, defaultNickname }) {
+  const { watch } = useFormContext();
+  const nickname = watch("nickname") ?? defaultNickname;
+
+  return mode === "new" ? (
+    <Trans
+      i18nKey="forms.signin_overlay.create_success_message"
+      components={[
+        <Styled.Heading />,
+        <Styled.Text />,
+        <Styled.Nickname as="h4" />
+      ]}
+      values={{ name: nickname }}
+    />
+  ) : (
+    <Styled.Heading>
+      <Trans
+        i18nKey="forms.signin_overlay.greeting"
+        components={[<Styled.Nickname />]}
+        values={{ name: nickname }}
+      />
+    </Styled.Heading>
+  );
+}
+
+ProfileGreeting.displayName = "Global.SignInUp.EditProfileForm.Greeting";
+
+ProfileGreeting.propTypes = {
+  mode: PropTypes.oneOf(["new", "existing"]),
+  defaultNickname: PropTypes.string
+};
