--- client/src/global/components/sign-in-up/Overlay/Overlay.js
@@ -1,20 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-import FocusTrap from "focus-trap-react";
-
-import { usePreventBodyScroll } from "hooks";
-import OverlayContent from "./OverlayContent";
-
-export default function Overlay({ hideOverlay }) {
-  usePreventBodyScroll();
-
-  return (
-    <FocusTrap
-      focusTrapOptions={{
-        onDeactivate: hideOverlay
-      }}
-    >
-      <OverlayContent hideOverlay={hideOverlay} />
-    </FocusTrap>
-  );
-}
