+++ client/src/global/components/sign-in-up/Overlay/index.js
@@ -1,1 +1,26 @@
-export { default } from "./Overlay";
+import React from "react";
+import PropTypes from "prop-types";
+import FocusTrap from "focus-trap-react";
+
+import { usePreventBodyScroll } from "hooks";
+import OverlayContent from "./OverlayContent";
+
+export default function Overlay({ hideOverlay }) {
+  usePreventBodyScroll();
+
+  return (
+    <FocusTrap
+      focusTrapOptions={{
+        onDeactivate: hideOverlay
+      }}
+    >
+      <OverlayContent hideOverlay={hideOverlay} />
+    </FocusTrap>
+  );
+}
+
+Overlay.displayName = "Global.SignInUp.Overlay";
+
+Overlay.propTypes = {
+  hideOverlay: PropTypes.func
+};
