+++ client/src/global/components/sign-in-up/form-inputs/Input/index.js
@@ -1,1 +1,45 @@
-export { default } from "./Input";
+import React, { forwardRef } from "react";
+import { useTranslation } from "react-i18next";
+import { useUID } from "react-uid";
+import * as Styled from "./styles";
+
+function Input(props, ref) {
+  const {
+    type = "text",
+    errors,
+    ariaLabel,
+    placeholder,
+    label,
+    name,
+    ...inputProps
+  } = props;
+
+  const { t } = useTranslation();
+  const uid = useUID();
+  const id = `${uid}_${name}`;
+  const errorId = `${name}_error`;
+
+  return (
+    <Styled.Field
+      name={`attributes.${name}`}
+      errors={errors}
+      idForError={errorId}
+    >
+      {label && <Styled.Label htmlFor={id}>{t(label)}</Styled.Label>}
+      <Styled.Input
+        id={id}
+        ref={ref}
+        type={type}
+        name={name}
+        aria-label={t(ariaLabel)}
+        aria-describedby={errorId}
+        placeholder={placeholder && t(placeholder)}
+        {...inputProps}
+      />
+    </Styled.Field>
+  );
+}
+
+Input.displayName = "Global.SignInUp.Inputs.Input";
+
+export default forwardRef(Input);
